% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{additiveGenotype}
\alias{additiveGenotype}
\title{Generate genotype calls for makeAdditive function}
\usage{
additiveGenotype(x, genotypeClasses, chromosome)
}
\arguments{
\item{x}{vector of two elements, [1]genotype call and [2]gender}

\item{genotypeClasses}{the different levels that the genotypes can manifest in}

\item{chromosome}{the chromosome of the genotype}
}
\value{
a list, with a model.frame object and a data.frame
}
\description{
Input an Genotype call, Chromosome and gender, returns a genotype call in numeric format
}
\examples{
gender<-c("Male","Female","Male","Female")
genotype<-c("A/A","A/B","A/A","B/B")

genotypeClasses<-c("A/A","A/B","B/B")

add<-additiveGenotype(cbind(genotype,gender)[1,],genotypeClasses,1)
identical(add,0)

add2<-apply( cbind( genotype , gender ) , 1 , additiveGenotype , genotypeClasses , 1 ) 
identical(add2,c(0,1,0,2))

}
