% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{alleleSplit}
\alias{alleleSplit}
\title{Split Alleles for MAF}
\usage{
alleleSplit(i, x, Chr = NULL, gender = NULL)
}
\arguments{
\item{i}{index of snp in X}

\item{x}{vector of genotypes for a single snp}

\item{Chr}{Chromosome each the snp is from}

\item{gender}{vector of the genders of each sample}
}
\value{
character vector of alleles
}
\description{
Split Alleles
}
\examples{
\dontrun{
  exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=4,byrow=TRUE)
 exampleChr<-1
 exampleGender<-c("M","M","F","M","F","F","F","F","M")

 identical(alleleSplit(1,exampleSNP[1,],exampleChr,exampleGender),c("A","A"))
 identical(alleleSplit(2,exampleSNP[1,],"X",exampleGender),c("A"))
}
}
