% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{callRate}
\alias{callRate}
\title{Calculate Call rate}
\usage{
callRate(x, crCutOff)
}
\arguments{
\item{x}{vector of genotypes}

\item{crCutOff}{cutoff for callrate}
}
\value{
vector of snps that failed to pass Call rate test, True means failure
}
\description{
Calculate call rates of the vector X, return true or false if meets call rate cutoff
}
\examples{
\dontrun{
  examplePASS<-c("A/A","A/B","A/A","A/B",NA,"B/B","A/A","A/A","B/B","A/A","A/B","A/A")
  exampleFAIL<-c(NA,NA,NA,NA,NA,NA,"A/A","B/B","B/B","A/A","B/B","B/B")

 callRate(examplePASS,.8)==FALSE
 callRate(exampleFAIL,.8)==TRUE
}
}
