% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{callRateIllumina}
\alias{callRateIllumina}
\title{Calculate Call rate for Illumina}
\usage{
callRateIllumina(x, crCutOff)
}
\arguments{
\item{x}{vector of genotypes}

\item{crCutOff}{cutoff for callrate}
}
\value{
vector of snps that failed to pass Call rate test, True means failure
}
\description{
Calculate call rates of the vector X, return true or false if meets call rate cutoff
}
\examples{
\dontrun{
  examplePASS<-c(1,2,1,2,NA,3,1,1,3,1,2,1)
  exampleFAIL<-c(NA,NA,NA,NA,NA,NA,1,3,3,1,3,3)

 callRateIllumina(examplePASS,.8)==FALSE
 callRateIllumina(exampleFAIL,.8)==TRUE
}
}
