% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{callRateTest}
\alias{callRateTest}
\title{Calculate Call rate -Master}
\usage{
callRateTest(x, crCutOff = 0.95, snowFall = FALSE)
}
\arguments{
\item{x}{Matrix of genotypes}

\item{crCutOff}{cutoff for callrate}

\item{snowFall}{Boolean value. determine whether to calculate callrate in series(FALSE) or Parallel(TRUE)}
}
\value{
vector of snps that failed to pass Call rate test
}
\description{
Master function to Calculate call rates of the snps in x, return true or false if fails to meets call rate cutoff
}
\examples{
\dontrun{
  exampleSNP<-matrix(c("A/A","A/B","A/A",NA,"A/B","B/B","A/A","A/A","B/B",NA,NA,NA,
 "A/A","A/B","A/A",NA,"A/A","B/B","B/B","A/A",NA,"B/B","B/B","A/A","A/A",
 "B/B",NA,"A/A",NA,"A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",NA,NA,
 "A/A",NA,"A/B","B/B","A/A","A/A","B/B",NA,"A/A","A/A",NA,"A/A","A/A",NA,"A/A","A/A"),nrow=8,byrow=TRUE)
 rownames(exampleSNP)<-c("snp1","snp2","snp3","snp4","snp5","snp6","snp7","snp8")

 all.equal(callRateTest(exampleSNP,.8),c("snp2"=2,"snp3"=3,"snp6"=6,"snp8"=8))
}

}
