% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{contrastGWAS}
\alias{contrastGWAS}
\title{Run a contrast on the lmObject}
\usage{
contrastGWAS(lmOut, con)
}
\arguments{
\item{lmOut}{LM object}

\item{con}{contrasts to perform. You can perform multiple contrasts in one. Takes the form: list(phenotype=list(pos=c("phenotypeLevel"),neg=c("phenotypeLevel")),...)}
}
\value{
a list object
}
\description{
Input an lm object and perform a contrast based on your input
}
\examples{
\dontrun{
examplePGx<-data.frame(var1=c(1,1,1,2,2,2,1,2),var2=c(2,2,2,3,3,3,3,2),var3=c(3,3,3,4,4,4,3,2),gender=c("Male","Female","Male","Female","Male","Female","Male","Male"))
exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B"),nrow=2,byrow=TRUE)
exampleSNPInfo<-data.frame(Chr=c(1,2))
exampleModel<-formula("var1~var2+var3+gender")
exampleCon<-list( "gender" = list(pos="Male",neg=c("Female")))

lmObject<-gwasFunc(1,exampleSNP,examplePGx,exampleModel,"var2","gender",databasePath="runGWAS.DB", dbTable="gwasFunc", snpInfo=exampleSNPInfo, overwriteEntry=TRUE)

contrastGWAS(lmObject$lm,exampleCon)
}
}
