% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{dumpFactorToCharacter}
\alias{dumpFactorToCharacter}
\title{Convert Factor to character}
\usage{
dumpFactorToCharacter(x)
}
\arguments{
\item{x}{A single character or factor instance}
}
\description{
convert factors to character for the covar file
}
\examples{
\dontrun{
 exampleDemographic<-data.frame(Phenotype1=c("Level1","Level1","Level2","Level2","Level1","Level2","Level1","Level1","Level2"),
                                Phenotype2=c(1,1,2,2,5,-1,2,-6,12),
                                Phenotype3=c("Level5","Leve2","Level2","Level3","Level10","Level1","Level7","Level1","Level1"),
                               SEX=c("F","M","M","F","F","F","M","M","M"),
                               row.names=c("Sub1","Sub2","Sub3","Sub4","Sub5","Sub6","Sub7","Sub8","Sub9"))
 
 identical(dumpFactorToCharacter(exampleDemographic["Sub1","Phenotype3"]),"Level5")
}
}
