% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModel.R
\name{findCorrelated}
\alias{findCorrelated}
\title{find correlated Phenotypes}
\usage{
findCorrelated(x, critVal = 0.7)
}
\arguments{
\item{x}{A dataframe of phenotypes, samples by row, phenotypes by column}

\item{critVal}{keep if phenotype in resulting correlation matric if it has a correlation with another phenotype greater than this value}
}
\value{
returns correlation matrix. keeps all phenotypes with a correlation above critVal
}
\description{
find correlated Phenotypes
}
\examples{
testDF<-data.frame(var1=c(1,1,1,1,2,2,2,2),var2=c(2,2,2,2,1,1,1,2),var3=c(2,3,2,4,3,4,2,2),var4=c(-1,-5,-2,-3,-4,-5,-1,-3))

findCorrelated(x = testDF, critVal=.7)
}
