% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModel.R
\name{findModel}
\alias{findModel}
\title{Create model from variables}
\usage{
findModel(x, depVar = NULL, indepVars, pCutOff = 0.1, exclude = NULL)
}
\arguments{
\item{x}{A dataframe of phenotypes, samples by row, phenotypes by column}

\item{depVar}{The name of the dependent variable in x. Defaults to NULL. If null, inputs random normalized values in for the depvar}

\item{indepVars}{Phenotypes to include in the model}

\item{pCutOff}{The pvalue cutoff to determine significant phenotypes}

\item{exclude}{Phenotypes to exclude from the model. defaults to NULL}
}
\value{
lm object
}
\description{
Create a model to use for the GWAS by running a linear model on phenotypes identified in indepvars without the snp component
}
\examples{
testDF<-data.frame(var1=c(1,1,1,1,2,2,2,2),var2=c(2,2,2,2,1,1,1,2),var3=c(2,3,2,4,3,4,2,2),var4=c(-1,-5,-2,-3,-4,-5,-1,-3))

newModel<-findModel(x = testDF,depVar = "var1",indepVars=c("var2","var3","var4"), pCutOff=0.05)

formula(newModel)==formula("var1~var2")

}
