% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{findMonomorphic}
\alias{findMonomorphic}
\title{Monomorphic SNP for Each SNP}
\usage{
findMonomorphic(x, snowFall = FALSE)
}
\arguments{
\item{x}{Matrix of genotypes}

\item{snowFall}{Boolean, should monomorphic calculations be performed in parallel(TRUE) or series(FALSE).}
}
\value{
named vector of monomorphic snps with index in x
}
\description{
Function to calculate if the SNP is monomorphic or not for each SNP in x
}
\examples{
\dontrun{
 exampleSNP<-matrix(c(0,1,0,1,2,0,0,2,0,1,0,0,2,2,0,2,2,0,0,2,0,1,0,0,2,2,0,2,0,1,0,1,2,0,0,2,0,0,0,0,0,0),nrow=7,byrow=TRUE)
 rownames(exampleSNP)<-c("snp1","snp2","snp3","snp4","snp5","snp6","snp7")
 
 findMonomorphic(exampleSNP)
}
}
