% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDiscordantCall}
\alias{getDiscordantCall}
\title{Get Discordant Genotypes}
\usage{
getDiscordantCall(x)
}
\arguments{
\item{x}{vector with 2 values}
}
\value{
data.frame
}
\description{
determine if calls are discordant or not
}
