% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getIBS}
\alias{getIBS}
\title{get IBS}
\usage{
getIBS(g, snowFall = FALSE)
}
\arguments{
\item{g}{matrix of genotypes, snp by row, sample by column}

\item{snowFall}{Boolean value for whether to perform this analysis in parallel or not. defaults to FALSE}
}
\value{
matrix of IBS values
}
\description{
calculate identical by state value
}
\examples{
\dontrun{
 exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=9,byrow=TRUE)
 rownames(exampleSNP)<-c(1,2,3,4,5,6,7,8,9)
 
 getIBS(exampleSNP,FALSE)
}
}
