% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getIBSNum}
\alias{getIBSNum}
\title{calculate IBS}
\usage{
getIBSNum(x)
}
\arguments{
\item{x}{vector containing 2 values}
}
\description{
calculates IBS number for two calls, either 1 or 0
}
