% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultRegressionFunctions.R
\name{gwasLMEHFuncsf}
\alias{gwasLMEHFuncsf}
\title{run a LMEH for a GWAS in Parallel}
\usage{
gwasLMEHFuncsf(i, g, pgx, Model, interactionTerm, genderVar, dbPath, dbTable,
  overwriteEntry = FALSE, tempFile = tempfile(), randomTerm,
  heterogeneousTerm)
}
\arguments{
\item{g}{vector object of numeric genotype calls for a single snp}

\item{pgx}{The data.frame containing the phenotypes used in the model}

\item{Model}{The original base model for the GWAS, excluding snp and snp interaction terms}

\item{interactionTerm}{The name of the variable to be included as the interaction term with the snp}

\item{dbPath}{The path to database to save results}

\item{overwriteEntry}{Boolean, passed to \code{\link[AxioSerializer]{writeObjectToTable}} whether to overwrite saved objects in the table specified}

\item{randomTerm}{The name of the random varable to include in results}

\item{heterogeneousTerm}{The name of the heterogeneous varable to include in results}
}
\value{
a P value, and saves lme models to the db specified in dbPath
}
\description{
A function called by runGWAS to perform a LMEH for each snp, designed for SnowFall parallel processing, and input snp is a  "CNSet" or "SnpSuperSet"
}
