% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultRegressionFunctions.R
\name{gwasMGLMFunc}
\alias{gwasMGLMFunc}
\title{run a MGLM for a GWAS}
\usage{
gwasMGLMFunc(i, g, pgx, Model, interactionTerm, genderVar, dbPath = NULL,
  dbTable = NULL, overwriteEntry = FALSE, tempFile = tempfile())
}
\arguments{
\item{i}{what row from SNP to call the genotypes from}

\item{g}{Matrix of "#/#" genotype calls, snp by row, subject by column}

\item{pgx}{The data.frame containing the phenotypes used in the model}

\item{Model}{The original base model for the GWAS, excluding snp and snp interaction terms}

\item{interactionTerm}{The name of the variable to be included as the interaction term with the snp}

\item{genderVar}{The name of the gender varable in pgx}

\item{dbPath}{The path to database to save results}

\item{dbTable}{The name of the table in the DB to save the results to}

\item{overwriteEntry}{Boolean, passed to \code{\link[AxioSerializer]{writeObjectToTable}} whether to overwrite saved objects in the table specified}

\item{family}{The family of Generalized Linear Model to use. Defaults to "binomial", more can be found at \code{\link[stats]{family}}}

\item{snpInfo}{Details about each snp. Requires at least a column titled 'Chr', for chromosome of snp}
}
\value{
list with 2 entries: pVal contains a P-Value, lm contains the linear model used
}
\description{
A function called by runGWAS to perform MGLM for each snp
}
\examples{
\dontrun{
examplePGx<-data.frame(var1=c(1,0,0,1,1,1,0,1),var2=c(2,2,2,3,3,3,3,2),var3=c(3,3,3,4,4,4,3,2),var4=c(1,2,1,1,2,1,1,2),var5=c(2,2,2,1,1,1,2,1),gender=c("Male","Female","Male","Female","Male","Female","Male","Male"))
exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B"),nrow=2,byrow=TRUE)
exampleSNPInfo<-data.frame(Chr=c(1,2))
exampleModel<-formula("var1~var2+var3+gender")

gwasLMEHFuncRes<-gwasMGLMFunc(i=1,g=exampleSNP, pgx=examplePGx,Model=exampleModel,interactionTerm="var2",genderVar="gender",family="gaussian",databasePath="runGWAS.DB", dbTable="gwasFunc", snpInfo=exampleSNPInfo, overwriteEntry=TRUE)
}
}
