% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultRegressionFunctions.R
\name{gwasMGLMFuncFF}
\alias{gwasMGLMFuncFF}
\title{run a MGLM for a GWAS in parallel or series}
\usage{
gwasMGLMFuncFF(i, g, pgx, Model, interactionTerm, genderVar, dbPath = NULL,
  dbTable = NULL, overwriteEntry = FALSE, tempFile = tempfile())
}
\arguments{
\item{g}{vector object of numeric genotype calls for a single snp}

\item{pgx}{The data.frame containing the phenotypes used in the model}

\item{Model}{The original base model for the GWAS, excluding snp and snp interaction terms}

\item{interactionTerm}{The name of the variable to be included as the interaction term with the snp}

\item{dbPath}{The path to database to save results}

\item{dbTable}{the name of the table to save the lm objects to. done using \code{\link[AxioSerializer]{writeObjectToTable}}}

\item{overwriteEntry}{Boolean, passed to \code{\link[AxioSerializer]{writeObjectToTable}} whether to overwrite saved objects in the table specified}

\item{family}{The family of Generalized Linear Model to use. Defaults to "binomial", more can be found at \code{\link[stats]{family}}}
}
\value{
list with 2 entries, a a P value, and the MGLM model. MGLM model is also saved to the db specified in databasePath.
}
\description{
A function called by runGWAS to perform a MGLM for each snp, designed for both SnowFall parallel processing and series.
}
