% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultRegressionFunctions.R
\name{gwasMGLMFuncsf}
\alias{gwasMGLMFuncsf}
\title{run a MGLM for a GWAS}
\usage{
gwasMGLMFuncsf(i, g, pgx, Model, interactionTerm, genderVar, dbPath, dbTable,
  overwriteEntry = FALSE, tempFile = tempfile())
}
\arguments{
\item{i}{index of snp of interest in g}

\item{g}{matrix object of numeric genotype calls, snp by row, sample by column}

\item{pgx}{The data.frame containing the phenotypes used in the model}

\item{Model}{The original base model for the GWAS, excluding snp and snp interaction terms}

\item{interactionTerm}{The name of the variable to be included as the interaction term with the snp}

\item{dbPath}{The path to database to save results}

\item{overwriteEntry}{Boolean, passed to \code{\link[AxioSerializer]{writeObjectToTable}} whether to overwrite saved objects in the table specified}

\item{family}{The family of Generalized Linear Model to use. Defaults to "binomial", more can be found at \code{\link[stats]{family}}}

\item{subCoef}{The name of the sub Coeffects varable to include in results}

\item{subEffect}{The name of the sub Effects varable to include in results}
}
\value{
list with 3 entries, a P value, dataframe of results and the name of the target. LM model is also saved to the db specified in dbPath
}
\description{
A helper function called by gwasMGLMFuncSF in runGWAS to perform a MGLM for each snp
}
