% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultRegressionFunctions.R
\name{gwasRRFuncFF}
\alias{gwasRRFuncFF}
\title{run a Ridge Regression for a GWAS in parallel or series}
\usage{
gwasRRFuncFF(i, g, pgx, Model, interactionTerm, genderVar,
  family = "binomial", dbPath = NULL, dbTable = NULL,
  overwriteEntry = FALSE, tempFile = tempfile())
}
\arguments{
\item{i}{the index of the snp in g to run the lm on}

\item{g}{matrix object of numeric genotype calls, snp by row, sample by column}

\item{pgx}{The data.frame containing the phenotypes used in the model}

\item{Model}{The original base model for the GWAS, excluding snp and snp interaction terms}

\item{interactionTerm}{The name of the variable to be included as the interaction term with the snp}

\item{genderVar}{The name of the genderVar in pgx}

\item{family}{The family of Generalized Linear Model to use. Defaults to "binomial", more can be found at \code{\link[stats]{family}}}

\item{dbPath}{The path to database to save results}

\item{dbTable}{the name of the table to save the lm objects to. done using \code{\link[AxioSerializer]{writeObjectToTable}}}

\item{overwriteEntry}{Boolean, passed to \code{\link[AxioSerializer]{writeObjectToTable}} whether to overwrite saved objects in the table specified}
}
\value{
list with 3 entries, a a P value, glm model, and ridge regression coefficients. glm model is also saved to the db specified in databasePath.
}
\description{
A function called by runGWAS to perform a GLM for each snp, designed for both SnowFall parallel processing and series. Adds interaction term using updateModelRR. Performs RidgeRegression.
}
