% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{hardyTest}
\alias{hardyTest}
\title{Hardy-Weinberg Equilibrium}
\usage{
hardyTest(x, returnAll = TRUE)
}
\arguments{
\item{x}{vector of genotypes for a single snp}

\item{returnAll}{Boolean value. If TRUE, returns the output of the fisher test. If FALSE, returns only the P-Values  from the fisher test.}
}
\description{
Calculate the Hardy-Weinberg Equilibrium for a SNP
}
\examples{
\dontrun{
 exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=2,byrow=TRUE)
 
 hardyTest(exampleSNP[1,],TRUE)
}
}
