% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{hardyTestIlluminaSF}
\alias{hardyTestIlluminaSF}
\title{Hardy-Weinberg Equilibrium for Illumina Calls}
\usage{
hardyTestIlluminaSF(i, x, returnAll = FALSE)
}
\arguments{
\item{x}{Vector of numeric (1,2,3) genotypes for a single SNP,}

\item{returnAll}{Boolean value. If TRUE, returns the output of the fisher test. If FALSE, returns only the P-Values  from the fisher test.}
}
\description{
Calculate the Hardy-Weinberg Equilibrium for a SNP from an Illumina Chip
}
\examples{
\dontrun{
 exampleSNP<-matrix(c(0,1,0,1,2,0,0,2,0,1,0,0,2,2,0,2,2,0,0,2,0,1,0,0,2,2,0,2,0,1,0,1,2,0,0,2),nrow=2,byrow=TRUE)
 
 hardyTestIllumina(exampleSNP[1,],TRUE)
}
}
