% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{hweTest}
\alias{hweTest}
\title{Hardy-Weinberg Equilibrium for ALL snps}
\usage{
hweTest(x, adjust = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
  "fdr", "none"), snowFall = FALSE)
}
\arguments{
\item{x}{Matrix of genotypes for a}

\item{adjust}{The method to adjust the P-values by after finding the HWE for each SNP. See \code{\link[stats]{p.adjust}} for more options}

\item{snowFall}{Boolean Value, should the calculations be performed in parallel(TRUE) or series(FALSE)}
}
\value{
vector of p-values
}
\description{
Calculate the Hardy-Weinberg Equilibrium for each SNP, and return adjusted P-Values
}
\examples{
\dontrun{
 exampleSNP<-matrix(
 c("A/A","A/A","A/A","A/A","A/A","A/A","A/A",
 "A/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A",
 "B/B","B/B","A/A","A/A","B/B","A/A","A/B","A/A",
 "A/A","B/B","B/B","A/A","B/B","A/A","A/B","A/A",
 "A/B","B/B","A/A","A/A","B/B","A/A","A/B","A/A",
 "A/B","B/B","A/A","A/A","B/B","A/A","A/B","A/A",
 "A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A",
 "B/B","A/A","A/B","B/B","A/A","A/A","A/B","B/B","A/A",
 ),nrow=8,byrow=TRUE)
 
 hweTest(exampleSNP)
}
}
