% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{mafFreq}
\alias{mafFreq}
\title{Calculate MAF}
\usage{
mafFreq(i, x, Chr = NULL, gender = NULL)
}
\arguments{
\item{i}{index of Matrix of X}

\item{x}{Matrix of genotypes}

\item{Chr}{vector of which chromosome each single snp is in}

\item{gender}{vector of the genders of the sample for each snp}
}
\value{
numeric Minor Allele Frequency
}
\description{
Calculate MAF for snp
}
\examples{
\dontrun{
  exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=4,byrow=TRUE)

 mafFreq(1,exampleSNP,c(1,1,2,2),c("M","M","F","M","F","F","F","F","M"))
}
}
