% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{mafFreqIllumina}
\alias{mafFreqIllumina}
\title{Calculate MAF for Illumina Data}
\usage{
mafFreqIllumina(x, Chr = NULL, gender = NULL)
}
\arguments{
\item{x}{vector of genotypes}

\item{Chr}{vector of which chromosome each single snp is in}

\item{gender}{vector of the genders of the sample for each snp}
}
\value{
numeric Minor allele frequency
}
\description{
Calculate MAF for snps from Illumina Chips
}
\examples{
\dontrun{
 exampleSNP<-c(0,1,0,1,2,0,0,2,0,0,1,0,1,2,0,0,2,0,0,1,0,1,2,0,0,2,0)

 mafFreqIllumina(exampleSNP,1,c("M","M","F","M","F","F","F","F","M"))
}
}
