% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{mafTest}
\alias{mafTest}
\title{test Minor Allele Frequencies}
\usage{
mafTest(x, qCutOff = 0.05, snowFall = FALSE, Chr = NULL, gender = NULL)
}
\arguments{
\item{x}{Matrix of genotypes}

\item{qCutOff}{cutoff for Minor Allele Frequency}

\item{snowFall}{Boolean Value, should the MAF calculations be performed in parallel(TRUE) or series(FALSE)}

\item{Chr}{vector of which chromosome each single snp is in}

\item{gender}{vector of the genders of the sample for each snp}
}
\value{
vector of snps that failed to pass MAF test
}
\description{
Calculate MAF for each snp, and identify SNPs failing to pass the Cutoff
}
\examples{
\dontrun{
  exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B","A/A","A/A","A/A","A/A"),nrow=4,byrow=TRUE)

 mafTest(exampleSNP, qCutOff = 0.5, snowFall=FALSE,Chr = c(1,1,2,2),gender = c("M","M","F","M","F","F","F","F","M","F"))
}
}
