% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{makeAdditive}
\alias{makeAdditive}
\title{Generate genotype calls for PGx data}
\usage{
makeAdditive(g, chromosome, gender)
}
\arguments{
\item{g}{vector of genotypes, In the format "#/#",}

\item{chromosome}{the chromosome the genotype is from}

\item{gender}{the vector of gender. Make sure the values are either "Male" or "Female"}
}
\value{
a vector of genotype calls numerically
}
\description{
Input an Genotype calls, Chromosome and gender, returns genotype calls in numeric format
}
\examples{
gender<-c("Male","Female","Male","Female")
genotype<-c("A/A","A/B","A/A","B/B")

add<-makeAdditive(genotype,1,gender)

identical(add,c(0,1,0,2))

}
