% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{makeCovarFile}
\alias{makeCovarFile}
\title{Create an Covariate file}
\usage{
makeCovarFile(file = "ped", covariates = c("GENDER"), demographics)
}
\arguments{
\item{file}{the filename of the Covariates file (will append a .cvr to the end)}

\item{covariates}{vector of phenotype names that is a column name in demographics to add to the covariates file}

\item{demographics}{data.frame containing phenotypes of the subjects. Samples by row, Phenotypes by column}
}
\description{
create the covariate file for Plink
}
\examples{
\dontrun{
 exampleDemographic<-data.frame(Phenotype1=c("Level1","Level1","Level2","Level2","Level1","Level2","Level1","Level1","Level2"),
                                Phenotype2=c(1,1,2,2,5,-1,2,-6,12),
                                Phenotype3=c("Level5","Leve2","Level2","Level3","Level10","Level1","Level7","Level1","Level1"),
                               SEX=c("F","M","M","F","F","F","M","M","M"),
                               row.names=c("Sub1","Sub2","Sub3","Sub4","Sub5","Sub6","Sub7","Sub8","Sub9"))
 
 makeAltPhenFile(file="test",covariates=c("Phenotype1","Phenotype2","Phenotype3"),demographics=exampleDemographic)
 
}
}
