% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{makeMapFile}
\alias{makeMapFile}
\title{Create MAP file for Plink}
\usage{
makeMapFile(file = "ped", phenotype = "ACSTRESN", snpInfo)
}
\arguments{
\item{file}{Name of the MAP file, will end up being 'file'-phenotype.map}

\item{phenotype}{The name of the phenotype to include in the map file name}

\item{snpInfo}{A data.frame of data about the SNPs, including a chr column and a pos column. snp names are rownames}
}
\description{
Converts SNP data and phenotype to MAP file for use in Plink
}
\examples{
\dontrun{
 exampleSNPInfo<-data.frame(chr=c("1","1","2","2","3","3","4","4","5","5"),
                            pos=c(1,2,1,2,1,2,1,2,1,2),
                            row.names=c("SNP1","SNP2","SNP3","SNP4","SNP5","SNP6","SNP7","SNP8","SNP9","SNP10"))
 
 makeMapFile(file="test",phenotype="Phenotype1",snpInfo=exampleSNPInfo)
}
}
