% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{makePedFile}
\alias{makePedFile}
\title{Create a PED file}
\usage{
makePedFile(file = "ped", phenotype = "ACSTRESN", demographics, genotypes)
}
\arguments{
\item{file}{the filename of the PED file}

\item{phenotype}{the categorical phenotype that is a column name in demographics}

\item{demographics}{data.frame containing phenotypes of the subjects. Samples by row, Phenotypes by column}

\item{genotypes}{an Object containing the genotypes. If not a CNset or SnpSuperSet, a matrix of Snp by row, subject by column}
}
\description{
create a PED file for Plink
}
\examples{
\dontrun{
 exampleDemographic<-data.frame(Phenotype1=c("Level1","Level1","Level2","Level2","Level1","Level2","Level1","Level1","Level2"),
                               SEX=c("F","M","M","F","F","F","M","M","M"),
                               row.names=c("Sub1","Sub2","Sub3","Sub4","Sub5","Sub6","Sub7","Sub8","Sub9"))
                               
 exampleGenotype<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=9,byrow=TRUE)
 rownames(exampleGenotype)<-c("Sub1","Sub2","Sub3","Sub4","Sub5","Sub6","Sub7","Sub8","Sub9")
 
 makePedFile(file="test",phenotype="Phenotype1",demographics=exampleDemographic,genotypes=exampleGenotype)
 
}
}
