% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{makePGxData}
\alias{makePGxData}
\title{Generate PGx data}
\usage{
makePGxData(x, g, chromosome, genderVar, ...)
}
\arguments{
\item{x}{data.frame of phenotypes}

\item{g}{vector of genotypes, In the format "#/#",}

\item{chromosome}{the chromosome the genotype is from}

\item{genderVar}{the colname of the gender variable in x. Make sure the values are either "Male" or "Female"}
}
\value{
a dataframe combining phenotype and genotype
}
\description{
Input an Phenotype DF, snpdata, chromosome and gender var and output a single dataframe
}
\examples{
testDF<-data.frame(var1=c(1,1,1,1),var2=c(2,2,2,2),var3=c(3,3,3,3),gender=c("Male","Female","Male","Female"))
genotype<-c("A/A","A/B","A/A","B/B")

pgx<-makePGxData(testDF,genotype,1,"gender")

identical(pgx,data.frame(testDF,Add=c(0,1,0,2)))

}
\seealso{
Other makePGxData: \code{\link{makePGxG2xData}}
}
