% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Plot Results of}
\usage{
manhattanPlot(x, type = "all", chr = NULL, xlab = "Chromosome",
  ylab = expression(-Log[10] ~ "p-Value"), main = "Manhattan Plot",
  reference = "hgu95av2", excludeXYM = TRUE, pCutOff = 0.05,
  sCutOff = 0.01, regions = NULL, rawChromosomes = FALSE, bleach = 0)
}
\arguments{
\item{x}{a dataframe containing the columns: CHR, pos, pValue. the rownames are the snpname}

\item{type}{?}

\item{chr}{vector of chromosomes to keep}
}
\description{
Converts SNP data and phenotype to MAP file for use in Plink
}
\examples{
\dontrun{
exampleX<-data.frame(CHR=c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3,4,4,4,4,4,4,4,4,4,4,4,4,5),
                     pos=c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,1,2,3,4,5,6,7,8,9,10,1,2,3,4,5,6,7,8,9,1,2,3,4,5,6,7,8,9,10,11,12,1))
exampleX$pValue<-runif(dim(exampleX)[1])
manhattanPlot(x=exampleX,chr=c(1,2,3,4))
}
}
