% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeDuplicates}
\alias{mergeDuplicates}
\title{Merge Duplicates}
\usage{
mergeDuplicates(x, g, uIndx = "PAT_ID", dIndx = "SPCM")
}
\arguments{
\item{x}{Sample Dataframe, phenotypes}

\item{g}{the function to peform the gwas}

\item{uIndx}{index of unique IDs}

\item{dIndx}{index of duplicated IDs}
}
\value{
genotype matrix
}
\description{
Merge duplicated values together
}
