% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeDuplicateVariable}
\alias{mergeDuplicateVariable}
\title{Merge Duplicated variables}
\usage{
mergeDuplicateVariable(x)
}
\arguments{
\item{x}{dataframe of variables}
}
\description{
Merge duplicated values together
}
