% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mergeGenotypes}
\alias{mergeGenotypes}
\title{Merge Duplicated Genotuypes}
\usage{
mergeGenotypes(x)
}
\arguments{
\item{x}{dataframe of variables}
}
\value{
data.frame
}
\description{
Merge duplicated values together
}
