% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{monomorphicSNP}
\alias{monomorphicSNP}
\title{Monomorphic SNP}
\usage{
monomorphicSNP(x)
}
\arguments{
\item{x}{vector of genotypes for a single snp}
}
\value{
boolean value
}
\description{
Calculate if the SNP is monomorphic or not
}
\examples{
\dontrun{
 exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=2,byrow=TRUE)
 
 monomorphicSNP(exampleSNP[1,])
}
}
