% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{monomorphicSNPIllumina}
\alias{monomorphicSNPIllumina}
\title{Monomorphic SNP for Illumina}
\usage{
monomorphicSNPIllumina(x)
}
\arguments{
\item{x}{vector of genotypes for a single snp}
}
\value{
boolean value
}
\description{
Calculate if the SNP is monomorphic or not from Illumina Chips
}
\examples{
\dontrun{
 exampleSNP<-matrix(c(0,1,0,1,2,0,0,2,0,1,0,0,2,2,0,2,2,0,0,2,0,1,0,0,2,2,0,2,0,1,0,1,2,0,0,2),nrow=2,byrow=TRUE)
 
 monomorphicSNPIllumina(exampleSNP[1,])
}
}
