% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{monomorphicSNPIlluminaSF}
\alias{monomorphicSNPIlluminaSF}
\title{Monomorphic SNP for Illumina, Snowfall}
\usage{
monomorphicSNPIlluminaSF(i, x)
}
\arguments{
\item{i}{index of Matrix to use in apply}

\item{x}{Matrix of genotypes, each row is a SNP, column is a sample}
}
\value{
vector of named boolean values
}
\description{
Helper function to calculate if the SNP is monomorphic or not from Illumina Chips, allow chunking by SnowFall package
}
\examples{
\dontrun{
 exampleSNP<-matrix(c(0,1,0,1,2,0,0,2,0,1,0,0,2,2,0,2,2,0,0,2,0,1,0,0,2,2,0,2,0,1,0,1,2,0,0,2),nrow=6,byrow=TRUE)
 rownames(exampleSNP)<-c("snp1","snp2","snp3","snp4","snp5","snp6")
 
 monomorphicSNPIlluminaSF(c(1:6),exampleSNP)
}
}
