% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotResults}
\alias{plotResults}
\title{Plot Results of}
\usage{
plotResults(x, pgx, genotypes, interactionTerm, pCutOff = 0.05, snpInfo,
  genderVar)
}
\arguments{
\item{x}{output from runGWAS of Pvalues}

\item{snpInfo}{A data.frame of data about the SNPs, including a chr column and a pos column. snp names are rownames}

\item{phenotype}{The name of the phenotype to include in the map file name}
}
\description{
Converts SNP data and phenotype to MAP file for use in Plink
}
\examples{
\dontrun{
examplePGx<-data.frame(var1=c(1,1,1,2,2,2,2,2,1,2,2),var2=c(2,2,2,3,3,3,2,2,3,2,3),var3=c(4,3,3,4,4,4,3,2,4,2,3),gender=c("Male","Female","Male","Female","Male","Female","Male","Male","Female","Female","Female"))
exampleSNP<-matrix(c("A/B","A/A","A/A","A/B","B/B","B/B","A/A","A/B","A/B","A/A","A/A",
                     "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B","A/A",
                     "A/B","B/B","A/A","A/A","B/B","A/A","A/B","A/B","A/B","B/B","A/A",
                     "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B"),
                     nrow=4,byrow=TRUE,dimnames=list(c("SNP1","SNP2","SNP3","SNP4"),NULL))
exampleSNPInfo<-data.frame(Chr=c(1,2,3,4))
exampleModel<-formula("var1~var2+var3+gender")

exampleX<-runGWAS(pgx=examplePGx,snp=exampleSNP,Model=exampleModel,func = AxioGWAS::gwasFunc,interactionTerm="var2",genderVar="gender",databasePath="runGWAS.DB", dbTable="runGWAS_Test", snpInfo=exampleSNPInfo, overwriteEntry=TRUE, adjust="none")

plotResults(x=exampleX,pgx=examplePGx,genotypes=exampleSNP,interactionTerm="var2", snpInfo = exampleSNPInfo, genderVar = "gender")
}
}
