% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{reducedModel}
\alias{reducedModel}
\title{Update model to include Add as variable of a function}
\usage{
reducedModel(x, Model)
}
\arguments{
\item{x}{data.frame where the colnames are found in the Model. Usually the phenotypes plus a column named "Add" for the genotypes of a snp (required)}

\item{Model}{the original formula to update. Must be a formula or model.frame (required)}
}
\value{
a model.frame object
}
\description{
Input an existing model, and will update the model with Add as a variable for the snp, strips interaction
}
\examples{
testDF<-data.frame(var1=c(1,1,1,1),var2=c(2,2,2,2),var3=c(3,3,3,3),var4=c(4,4,4,4),Add=c(0,0,0,0))
Model<-formula("var1~var2+var3+var4")

newModel<-reduceModel(testDF,Model)

identical(formula(newModel),formula("var1~var2+var3+Add"))
}
