% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{runContrastGWAS}
\alias{runContrastGWAS}
\title{Run a contrast on the results of the GWAS}
\usage{
runContrastGWAS(i, func, fd, pgx, Model, interactionTerm, genderVar, con, ...)
}
\arguments{
\item{i}{index of of snp in fd}

\item{func}{the function to peform the gwas}

\item{fd}{Matrix of genotype calls, snp by row, subject by column. Genotype format varies by called "func".}

\item{pgx}{DF of phenotypes for each subject}

\item{Model}{the original formula to update. Must be a formula or model.frame (required)}

\item{interactionTerm}{One of the colnames of x, to be added to the model, and interact with the genotype}

\item{genderVar}{The name of the gender varable in pgx}

\item{con}{contrasts to perform. You can perform multiple contrasts in one. Takes the form: list(phenotype=list(pos=c("phenotypeLevel"),neg=c("phenotypeLevel")),...)}

\item{...}{additional arguments to be passed to func}
}
\value{
a list object
}
\description{
Input one of the gwas functions, and it will perform contrasts and return a list with entries for estOut and pValue
}
\examples{
\dontrun{
examplePGx<-data.frame(var1=c(1,1,1,2,2,2,1,2),var2=c(2,2,2,3,3,3,3,2),var3=c(3,3,3,4,4,4,3,2),gender=c("Male","Female","Male","Female","Male","Female","Male","Male"))
exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B"),nrow=2,byrow=TRUE)
exampleSNPInfo<-data.frame(Chr=c(1,2))
exampleModel<-formula("var1~var2+var3+gender")
exampleCon<-list( "gender:var2" = list(pos="Male/2",neg=c("Male/3","Female/2","Female/3")))

newContrast<-runContrastGWAS(1,gwasFunc,exampleSNP,examplePGx,exampleModel,"var2","gender",exampleCon,
               databasePath="runGWAS.DB", dbTable="gwasFunc", snpInfo=exampleSNPInfo, overwriteEntry=TRUE)
}
}
