% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpFilter.R
\name{runsnpFilter}
\alias{runsnpFilter}
\title{Run snpFilter}
\usage{
runsnpFilter(x, crCutOff, mafCutOff, hweCutOff, Chr, gender,
  adjust = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
  "none"), snowFall = FALSE, returnAll = FALSE)
}
\arguments{
\item{x}{vector of genotypes in "#/#" format}

\item{crCutOff}{Call Rate CutOff}

\item{mafCutOff}{Minor allele Frequency Cutoff}

\item{hweCutOff}{Hardy-Weinberg Equilibrium P-Value Cutoff}

\item{Chr}{named vector of which chromosome each single snp is in}

\item{gender}{vector of the genders of the sample for each snp}

\item{adjust}{pvalue adjustment method}

\item{snowFall}{Perform testing in parallel(TRUE) or in series(FALSE)}

\item{returnAll}{return HWE pvalues for all passing SNP}
}
\value{
vector of indeces of passing SNP
}
\description{
Identify if a snp passes All Filtering Steps
}
\examples{
\dontrun{
  exampleGender<-c("Male","Male","Female","Female","Male","Male","Female","Female","Male","Male","Female")
  
  exampleSNP1<-c("A/A","A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B")#callrate=100\%,MAF=.4545, will return a HWE value
  exampleSNP2<-c("A/A","A/A","A/A","A/A","A/A","A/A","A/A","A/A","A/A","A/A","A/B")#callrate=100\%,MAF=.04545, will fail MAF
  exampleSNP3<-c(NA,NA,NA,"A/A","A/A","A/A","A/A","A/A","A/A","A/A","A/B")#callrate=72.72\%, will fail callRate
  exampleSNP4<-c("A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B","A/B")#callrate=100\%,MAF=.5, will return a HWE value
  exampleSNPMatrix<-matrix(c(exampleSNP1,exampleSNP2,exampleSNP3,exampleSNP4),nrow=4,byrow=TRUE,dimnames=list(c("SNP1","SNP2","SNP3","SNP4")))
  
  exampleChr<-c("SNP1"=1,"SNP2"=1,"SNP3"=1,"SNP4"=1)
  
  runsnpFilter(exampleSNPMatrix,crCutOff=.95,mafCutOff=.05, hweCutOff=0.05,Chr=exampleChr,gender=exampleGender,returnAll=TRUE)
  KeepSNPS<-runsnpFilter(exampleSNPMatrix,crCutOff=.95,mafCutOff=.05, hweCutOff=0.05,Chr=exampleChr,gender=exampleGender,returnAll=FALSE)
  
  #make exampleSNPMatrix into a SnpSuperSet
  exampleSNPMatrix<-matrix(as.numeric(gsub("A/B",1,gsub("A/A",0,exampleSNPMatrix))),nrow=4,byrow=FALSE,dimnames=list(c("SNP1","SNP2","SNP3","SNP4")))
  exampleSNPSS<-new("SnpSuperSet",alleleA=exampleSNPMatrix,alleleB=exampleSNPMatrix,call=exampleSNPMatrix,callProbability=exampleSNPMatrix)
  runsnpFilter(exampleSNPSS,crCutOff=.95,mafCutOff=.05, hweCutOff=0.05,Chr=exampleChr,gender=exampleGender,returnAll=TRUE)
  KeepSNPS<-runsnpFilter(exampleSNPSS,crCutOff=.95,mafCutOff=.05, hweCutOff=0.05,Chr=exampleChr,gender=exampleGender,returnAll=FALSE)
  
  
}
}
