% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scaleXAxis}
\alias{scaleXAxis}
\title{Plot Results of}
\usage{
scaleXAxis(x, scale = FALSE, reference = "hgu95av2", excludeXYM = TRUE)
}
\arguments{
\item{x}{dataframe containing chromosome and position for the snps intended to use}

\item{scale}{The name of the phenotype to include in the map file name}

\item{reference}{Which genome to use as reference}

\item{excludeXYM}{Boolean value, whether to keep (FALSE) or remove (TRUE) the X,Y and M Chromosomes}
}
\description{
Converts SNP data and phenotype to MAP file for use in Plink
}
\examples{
\dontrun{
exampleX<-data.frame(CHR=c(1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2,3,3,3,3,3,3,3,3,3,4,4,4,4,4,4,4,4,4,4,4,4,5),
                     pos=c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,1,2,3,4,5,6,7,8,9,10,1,2,3,4,5,6,7,8,9,1,2,3,4,5,6,7,8,9,10,11,12,1))
unscaledXAxis<-scaleXAxis(x=exampleX,scale=FALSE,reference="hgu95av2",excludeXYM=TRUE)
scaledXAxis<-scaleXAxis(x=exampleX,scale=TRUE,reference="hgu95av2",excludeXYM=TRUE)
}
}
