% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpFilter.R
\name{snpFilterIllumina}
\alias{snpFilterIllumina}
\title{snp Filter Illumina}
\usage{
snpFilterIllumina(i, x, crCutOff, mafCutOff, Chr, gender)
}
\arguments{
\item{i}{index of snp in x to apply snpFilter on}

\item{x}{Matrix of genotypes in "#/#" format}

\item{crCutOff}{Call Rate CutOff}

\item{mafCutOff}{Minor allele Frequency Cutoff}

\item{Chr}{vector of which chromosome each single snp is in}

\item{gender}{vector of the genders of the sample for each snp}
}
\value{
dataframe of SNP index and HWE pvalue
}
\description{
Determine if snp passes Filtering Steps (MAF, CallRate), if so, return HWE pvalue, for Illumina
}
\examples{
\dontrun{
  exampleGender<-c("Male","Male","Female","Female","Male","Male","Female","Female","Male","Male","Female")
  exampleSNP1<-c(0,1,1,1,1,1,1,1,1,1,1)#callrate=100\%,MAF=.4545, will return a HWE value
  exampleSNP2<-c(0,0,0,0,0,0,0,0,0,0,1)#callrate=100\%,MAF=.04545, will fail MAF
  exampleSNP3<-c(NA,NA,NA,0,0,0,0,0,0,0,1)#callrate=72.72\%, will fail callRate#'   
  exampleSNPMatrix<-matrix(c(exampleSNP1,exampleSNP2,exampleSNP3),nrow=3,byrow=TRUE,dimnames=list(c("SNP1","SNP2","SNP3")))
  
  exampleChr<-c(1,1,1)
  
  snpFilterIllumina(1,exampleSNPMatrix,crCutOff=.95,mafCutOff=.05,Chr=exampleChr,gender=exampleGender)
  
  snpFilterIllumina(2,exampleSNPMatrix,crCutOff=.95,mafCutOff=.05,Chr=exampleChr,gender=exampleGender)
  
  snpFilterIllumina(3,exampleSNPMatrix,crCutOff=.95,mafCutOff=.05,Chr=exampleChr,gender=exampleGender)
  
}
}
