% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{stripGLM_callEnv}
\alias{stripGLM_callEnv}
\title{Clean LM objects of all kinds results and store in table}
\usage{
stripGLM_callEnv(LMINPUT)
}
\arguments{
\item{LMINPUT}{any sort of lm object}
}
\value{
stripped down lm object
}
\description{
A helper function to reduce the size of the lm objects. They had connections to the environment the lm object was created in, and would become large when serialized
}
