% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summaryIBSMat}
\alias{summaryIBSMat}
\title{get IBS}
\usage{
summaryIBSMat(x, cutOff = 0.95)
}
\arguments{
\item{x}{matrix of IBS values from \code{getIBS}}

\item{cutOff}{Level at which to identify subjects with IBS > cutoff}
}
\value{
matrix of IBS values
}
\description{
calculate identical by state value
}
\examples{
 exampleSNP<-matrix(c("A/A","A/B","A/A","A/B","B/B","A/A","A/A","B/B",
 "A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","B/B","A/A","A/A",
 "B/B","A/A","A/B","A/A","A/A","B/B","B/B","A/A","B/B","A/A","A/B",
 "A/A","A/B","B/B","A/A","A/A","B/B"),nrow=6,byrow=TRUE)
 rownames(exampleSNP)<-c(1,2,3,4,5,6)
 
 ibsMat<-getIBS(exampleSNP,FALSE)

 summaryIBSMat(ibsMat,.75)
 
}
