% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{updateModel}
\alias{updateModel}
\title{Update model to include interaction terms or replace parts of a function}
\usage{
updateModel(x, Model, interactionTerm = NULL, replace = NULL)
}
\arguments{
\item{x}{data.frame where the colnames are found in the Model. Usually the phenotypes plus the genotypes of a snp (required)}

\item{Model}{the original formula to update. Must be a formula or model.frame (required)}

\item{interactionTerm}{One of the colnames of x, to be added to the model, and interact with the genotype}

\item{replace}{Remove the specified term from the model}
}
\value{
a model.frame object
}
\description{
Input an existing model, and will update the model based on your inputs
}
\examples{
testDF<-data.frame(var1=c(1,1,1,1),var2=c(2,2,2,2),var3=c(3,3,3,3),var4=c(4,4,4,4),Add=c(0,0,0,0))
Model<-formula("var1~var2+var3+var4")

newModel<-updateModel(testDF,Model,interaction="var2",replace="var4")

identical(formula(newModel),formula("var1~var2+var3+Add+var2:Add"))

}
