% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{updateModelRR}
\alias{updateModelRR}
\title{Update model to include interaction terms or replace parts of a function for RR}
\usage{
updateModelRR(x, Model, interactionTerm = NULL, replace = NULL)
}
\arguments{
\item{x}{data.frame where the colnames are found in the Model. Usually the phenotypes plus a column named "Add" for the genotypes of a snp (required)}

\item{Model}{the original formula to update. Must be a formula or model.frame (required)}

\item{interactionTerm}{One of the colnames of x, to be added to the model, and interact with the genotype}

\item{replace}{Remove the specified term from the model}
}
\value{
a list, with a model.frame object and a data.frame
}
\description{
Input an existing model, and will update the model for RR based on your inputs
}
\examples{
testDF<-data.frame(var1=c(1,1,1,1),var2=c(2,2,2,2),var3=c(3,3,3,3),var4=c(4,4,4,4),Add=c(0,0,0,0))
Model<-formula("var1~var2+var3+var4")

newModel<-updateModelRR(testDF,Model,interactionTerm="var2",replace="var4")

identical(formula(newModel$newForm),formula("var1~var2+var3+Add+var2:Add"))
}
